+++ R/install.R
@@ -4,13 +4,13 @@
 #' direct dependencies. For a more complete installation that includes
 #' additional optional dependencies, use [`keras::install_keras()`].
 #'
-#' @details You may be prompted to download and install
-#'   miniconda if reticulate did not find a non-system installation of python.
-#'   Miniconda is the recommended installation method for most users, as it
-#'   ensures that the R python installation is isolated from other python
-#'   installations. All python packages will by default be installed into a
-#'   self-contained conda or venv environment named "r-reticulate". Note that
-#'   "conda" is the only supported method on M1 Mac.
+#' @details You may be prompted to download and install miniconda if reticulate
+#'   did not find a non-system installation of python. Miniconda is the
+#'   recommended installation method for most users, as it ensures that the R
+#'   python installation is isolated from other python installations. All python
+#'   packages will by default be installed into a self-contained conda or venv
+#'   environment named "r-reticulate". Note that "conda" is the only supported
+#'   method on M1 Mac.
 #'
 #'   If you initially declined the miniconda installation prompt, you can later
 #'   manually install miniconda by running [`reticulate::install_miniconda()`].
@@ -85,15 +85,17 @@
 #' @param python_version,conda_python_version Pass a string like "3.8" to
 #'   request that conda install a specific Python version. This is ignored when
 #'   attempting to install in a Python virtual environment. Note that the Python
-#'   version must be compatible with the requested Tensorflow version, documented
-#'   here: <https://www.tensorflow.org/install/pip#system-requirements>
+#'   version must be compatible with the requested Tensorflow version,
+#'   documented here:
+#'   <https://www.tensorflow.org/install/pip#system-requirements>
 #'
 #' @param pip_ignore_installed Whether pip should ignore installed python
 #'   packages and reinstall all already installed python packages. This defaults
 #'   to `TRUE`, to ensure that TensorFlow dependencies like NumPy are compatible
 #'   with the prebuilt TensorFlow binaries.
 #'
-#' @param fresh If `TRUE`, any existing environment specified by `envname` is deleted first.
+#' @param new_env If `TRUE`, any existing Python virtual environment and/or
+#'   conda environment specified by `envname` is deleted first.
 #'
 #' @param ... other arguments passed to [`reticulate::conda_install()`] or
 #'   [`reticulate::virtualenv_install()`], depending on the `method` used.
@@ -113,7 +115,7 @@ function(method = c("auto", "virtualenv", "conda"),
          conda_python_version = NULL,
          ...,
          pip_ignore_installed = TRUE,
-         fresh = identical(envname, "r-tensorflow"),
+         new_env = identical(envname, "r-tensorflow"),
          python_version = NULL) {
 
   method <- match.arg(method)
@@ -146,7 +148,7 @@ function(method = c("auto", "virtualenv", "conda"),
     as.character(extra_packages)
   ))
 
-  if (isTRUE(fresh)) {
+  if (isTRUE(new_env)) {
 
     if (method %in% c("auto", "virtualenv") &&
         reticulate::virtualenv_exists(envname))