+++ R/install.R
@@ -25,16 +25,12 @@
 #'   Sys.setenv("RETICULATE_PYTHON" = "~/path/to/python-env/bin/python")
 #'   ```
 #'
-#' @section Apple Silicon: Tensorflow on Apple Silicon is not officially
-#'   supported by the Tensorflow maintainers. However Apple has published a
-#'   custom version of Tensorflow compatible with Arm Macs.
-#'   `install_tensorflow()` will install the special packages `tensorflow-macos`
-#'   and `tensorflow-metal` on Arm Macs. See
+#' @section Apple Silicon: Beginning with Tensorflow version 2.13, the default
+#'   tensorflow package now works on Apple Silicon. See
 #'   \url{https://developer.apple.com/metal/tensorflow-plugin/} for instructions
-#'   on how to do the equivalent manually. Please note that this is an
-#'   experimental build of both Python and Tensorflow, with known issues. In
-#'   particular, certain operations will cause errors, but can often be remedied
-#'   by pinning them to the CPU. For example:
+#'   on how to install older versions of Tensorflow on macOS. Please note that
+#'   not all operations are supported on Arm Mac GPUs. You can work around the
+#'   missing operations by pinning operations to CPU. For example:
 #'
 #'   ```` R
 #'   x <- array(runif(64*64), c(1, 64, 64))
@@ -232,7 +228,6 @@ function(method = c("auto", "virtualenv", "conda"),
     msg <- "# Configured by the R function tensorflow::install_tensorflow()"
 
     if(!all(c(msg, vars) %in% profile)) {
-      browser()
       profile <- c(profile, "", msg, vars, "")
       writeLines(profile, "~/.profile")
       message("- Updated file '~/.profile' with new environment variables.")