+++ NEWS.md
@@ -1,3 +1,5 @@
+# tensorflow (development version)
+
 # tensorflow 2.11.0
 
 - `install_tensorflow()` now installs TF v2.11 by default.